KH Classify: extract Tag Library tab into ClassifyTagsTabComponent
Tag table, search, word cloud, tags-by-source, and drill-down moved to a
self-contained sub-page. Deep-link tag search + SearchClassifyTags agent tool
delegate to the tab via @ViewChild ApplySearch(). Item detail bubbles up via
OpenItemDetailRequested (slide-in stays host-owned). Build green.
