+++ packages/Angular/Explorer/dashboards/src/AI/components/autotagging/tabs/tags-tab.component.css
@@ -1,0 +1,8 @@
+/*
+ * Tag Library tab styles.
+ *
+ * The host (AutotaggingPipelineResourceComponent) uses ViewEncapsulation.None,
+ * so its `.at-*` rules are global and already style this tab's markup while it
+ * renders inside the host shell. Tab-specific rules live here as the shared CSS
+ * is partitioned during the host-slim step.
+ */
